beta-issues/requirements.txt
appdirs==1.4.3
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
passlib==1.7.1
pdfdocument==3.2
pdfminer.six==20181108
pycryptodome==3.7.2
pyparsing==2.3.0
six==1.11.0
sortedcontainers==2.1.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
WTForms==2.2.1
